# docs/swagger.yaml
basePath: /
definitions:
...
  api.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  api.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.registerRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  httpx.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorDetail'
    type: object
  httpx.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      xp:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Backend for the Jigsaw Puzzle app: email/password auth, image
    collection, admin daily challenges (Cloudinary) and per-day
    reputation ("Danh vọng").
  title: Jigsaw Puzzle API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates with email + password and returns a session JWT.
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      summary: Login
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates an account with email + password (password is bcrypt-hashed)
        and returns a session JWT.
      parameters:
      - description: Registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      summary: Register
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token issued by
      /api/v1/auth/login or /register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"